Better godef integration: local defs, imports
go-plus' current godef integration is limited to top-level and exported names,
and does not work for method calls:

    buf := bytes.Buffer{}
    buf.Write([]byte{0x0}) // godef on Write and buf didn't work.

This is because godef was being run as:

    godef -f filebeingedited.go buf.Write

which doesn't give godef enough context to know what "buf" and "Write" are.

The fix is to run godef with the offset (-o) option to give it the offset
in the file where the cursor is, and let godef figure out what the user's
trying to navigate to.

This commit also keeps track of godef navigation on a stack and allows going
back to previous locations where godef was triggered with the godef-return
command.