Support godoc.org links for third party package doc
The godoc.org link feature requires an update to gogetdoc in order to work properly when the cursor is over the import clause of a third-party package.  Specifically, gogetdoc needs to provide the full import path in all cases.  It currently omits the import path for package documentation.

This commit adds support for the (yet to be released) gogetdoc update, which will enable to godoc.org links to work correctly for third party package doc.

In addition, it adds logic to detect when this gogetdoc update has not yet been installed, and prompts the user to update.  The user's decision is remembered so that they aren't repeatedly asked to update.