Add save event orchestrator
This will be a one stop shop for subscribing to on{Will|Did}Save events.

Currently we only support `onWillSave` events.  Support for `onDidSave` is forthcoming.

The orchestrator executes callbacks in the order they are registered,
and allows a callback to prevent execution of future callbacks.

Convert formatter over to using the orchestrator rather than subscribing manually.