Ensure that usage panel displays progress when focused
Since the view gets destroyed and recreated each time the tab focuses, it is often the case that usage.view is undefined when updating content.  This causes us to skip an update so the "Running guru" content never gets rendered.

The fix is to make requestFocus() return a promise that resolves when the view is created and initially rendered.  We can chain off this promise for future updates to ensure that the view is defined.