Make filenames in test output panel links
This is a first cut at the feature.

Implement a <FileLinks> etch component that takes arbitrary text and turns anything that matches a regex into a link with a configurable click handler.

This approach may not be valid, as it currently breaks colored output from ansi2html (we expect raw text and not HTML).  Perhaps a similar component that handles ANSI escape codes could replace our use of ansi2html.
Note: we can't use the innerHTML property anymore because it wonder render JSX.