Remove usage of ansi-to-html
Restore support for colored test output.

Instead of using ansi-to-html to build a raw HTML string, we're now using ansi-style-parser to build etch components that are styled with less.

Now that we have real components and not an opaque string we should have an easier time adding <a> elements to link to files.

Also fixes #555 (white space not respected in test output)