Add file link functionality
Add an optional mapper func property to the <Ansi> element that will be called with each element produced from the ansi parser.

If no mapper is provided, the default one just puts all text in a styled span.

Include a mapper that looks for file locations (*.go:##) and generates links to them.

Note: this doesn't play nice with the bogus file paths we get due to running with coverage.  (They're deleted by the time tests finish and the link is broken)