Ensure that we generate links
Fixes a bug where links arent generated if the tab hasn't been opened yet.

This is another case where the EtchComponent base class hurts us.  It guarantees that etch.initialize (and in turn our component's update method) gets called before the constructor completes.

As a result, the class is not fully initialized.  In this case, the regex hasn't been assigned, so we can't use it to detect file positions.

The solution is to move the regex outside of the class so that it is always valid.