add option to disable compilation on save
Add option `go-plus.config.compileOnSave` that defaults to true.

When disabled, go plus will not attempt to build your code when a file is saved.
In this mode, go-plus will only format the file on save, which can be helpful if for some reason local compilation would fail (see #674).

Additionally, we make the assumption that if the user doesn't want to compile on save, they also don't want to test on save (if local compilation would fail, local test runs would also fail), and so we never run tests on save if compile on save is disabled.