Add resize-observer to hide tab labels
We no longer focus on differentiation when the panel is _vertical_.  Instead, we focus on when the panel is _narrow_.  This allows us to hide the tab labels when the panel is docked on the left or right sides, or when docked in the center alongside another element.

For now we use a resize-observer polyfill - this can be removed once the feature is native to Chrome (it's currently hidden behind an experimental flag.