simplify executor options
The previous behavior of allowing a string or an object for the getOptions call was confusing and error prone.

getOptions was used as both a public API call to return a sensible set of default options for a given type (project or file), AND as an internal call to ensure that we fill in any required options that the user didn't specify.

By separating these two behaviors we can simplify the public API and make things much more manageable.