get rid of fsevents dependency
luckily fsevents is only a dependecy of babel-plugin-add-module-exports.
This plugin transforms default exports into something that is easier to
consume within nodejs.
The fix is rather easy: "just" get rid of default exports.
Unfortunately this was tricker than initially thought because several
files did not use flowtype and the main file did not type check
properly.